+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- ...
+- `fiboa rename-extension` to quickly replace template placeholders in new extensions
 
 ### Changed
 
@@ -26,6 +26,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - ...
+
+## [v0.3.2] - 2024-04-12
+
+### Added
+
+- `fiboa rename-extension` to quickly replace template placeholders in new extensions
 
 ## [v0.3.1] - 2024-04-11
 
@@ -130,7 +136,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - First release
 
-[Unreleased]: <https://github.com/fiboa/cli/compare/v0.3.1...main>
+[Unreleased]: <https://github.com/fiboa/cli/compare/v0.3.2...main>
+[v0.3.2]: <https://github.com/fiboa/cli/compare/v0.3.1...v0.3.2>
 [v0.3.1]: <https://github.com/fiboa/cli/compare/v0.3.0...v0.3.1>
 [v0.3.0]: <https://github.com/fiboa/cli/compare/v0.2.1...v0.3.0>
 [v0.2.1]: <https://github.com/fiboa/cli/compare/v0.2.0...v0.2.1>